# backend/openapi/deadmans.v1.yaml
openapi: 3.0.3
info:
  title: Dead-Mans API
  version: v1
  description: >
    Canonical transport contract for Dead-Mans: HTTP paths under /api and SignalR hubs
    documented in x-signalr. Regenerate frontend contracts after changes.
servers:
  - url: /api
paths:
...
  /integrations/twitch/status:
    get:
      operationId: getTwitchQuizIntegrationStatus
      security:
        - cookieAuth: []
      responses:
        '200':
          description: Twitch quiz integration connection and delivery state
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TwitchQuizIntegrationStatusDto'
  /integrations/twitch/oauth/{role}:
    get:
      operationId: connectTwitchQuizAccount
      security:
        - cookieAuth: []
      parameters:
        - name: role
          in: path
          required: true
          schema:
            type: string
            enum: [bot, broadcaster]
      responses:
        '302':
          description: Redirect to Twitch authorization
  /integrations/twitch/oauth/callback:
    get:
      operationId: completeTwitchQuizOAuth
      security:
        - cookieAuth: []
      parameters:
        - name: code
          in: query
          required: true
          schema: { type: string }
        - name: state
          in: query
          required: true
          schema: { type: string }
      responses:
        '302':
          description: OAuth grant stored and browser redirected to the quiz panel
...
  /integrations/twitch/quiz/questions/prepare:
    post:
      operationId: prepareTwitchQuizQuestion
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PrepareTwitchQuestionRequestDto'
      responses:
        '202':
          description: Durable publication created; timer has not started yet
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PrepareTwitchQuizQuestionResultDto'
        '400':
          description: Integration disabled, disconnected, or question incompatible
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '404':
          description: No active game or available question
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '409':
          description: Publication or unresolved outcome already exists
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
  /integrations/twitch/quiz/publications/{publicationId}/retry:
    post:
      operationId: retryTwitchQuizPublication
      security:
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/TwitchPublicationId'
      responses:
        '200':
          description: Failed or uncertain step explicitly queued again
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TwitchQuizPublicationStateDto' }
        '404': { description: Publication not found }
  /integrations/twitch/quiz/publications/{publicationId}/cancel:
    post:
      operationId: cancelTwitchQuizPublication
      security:
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/TwitchPublicationId'
      responses:
        '200':
          description: Preparation cancelled; a published question receives a cancellation message
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TwitchQuizPublicationStateDto' }
        '404': { description: Publication not found }
  /integrations/twitch/quiz/publications/{publicationId}/skip-outcome:
    post:
      operationId: skipTwitchQuizOutcome
      security:
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/TwitchPublicationId'
      responses:
        '200':
          description: Unresolved result delivery explicitly skipped
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TwitchQuizPublicationStateDto' }
        '404': { description: Publication not found }
...
components:
  parameters:
    TwitchPublicationId:
      name: publicationId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: dm_auth
  schemas:
...
    TwitchQuizPublicationStateDto:
      type: object
      required: [publicationId, gameId, questionId, askOrder, status, questionDeliveryStatus, optionsDeliveryStatus, outcomeDeliveryStatus, questionMessage, optionsMessage, createdAtUtc, updatedAtUtc]
      properties:
        publicationId: { type: string, format: uuid }
        gameId: { type: string, format: uuid }
        questionId: { type: string, format: uuid }
        questionSessionId: { type: string, format: uuid, nullable: true }
        askOrder: { type: integer, minimum: 1 }
        status: { type: string, enum: [publishing, open, failed, uncertain, cancel_pending, cancelled, completed] }
        questionDeliveryStatus: { type: string, enum: [pending, sending, sent, failed, uncertain, skipped] }
        optionsDeliveryStatus: { type: string, enum: [pending, sending, sent, failed, uncertain, skipped] }
        outcomeDeliveryStatus: { type: string, enum: [pending, sending, sent, failed, uncertain, skipped] }
        questionMessage: { type: string, maxLength: 500 }
        optionsMessage: { type: string, maxLength: 500 }
        lastError: { type: string, nullable: true }
        createdAtUtc: { type: string, format: date-time }
        updatedAtUtc: { type: string, format: date-time }
    TwitchQuizIntegrationStatusDto:
      type: object
      required: [enabled, botConnected, broadcasterConnected, eventSubConnected, accessRevoked]
      properties:
        enabled: { type: boolean }
        botConnected: { type: boolean }
        broadcasterConnected: { type: boolean }
        eventSubConnected: { type: boolean }
        accessRevoked: { type: boolean }
        botUserId: { type: string, nullable: true }
        broadcasterUserId: { type: string, nullable: true }
        lastError: { type: string, nullable: true }
        publication:
          type: object
          allOf:
            - $ref: '#/components/schemas/TwitchQuizPublicationStateDto'
          nullable: true
    PrepareTwitchQuestionRequestDto:
      type: object
      properties:
        questionId: { type: string, format: uuid, nullable: true }
    PrepareTwitchQuizQuestionResultDto:
      type: object
      required: [outcome]
      properties:
        outcome:
          type: string
          enum: [prepared, disabled, notConnected, noActiveGame, noAvailableQuestions, publicationInProgress, pendingOutcome, incompatibleQuestion]
        publication:
          type: object
          allOf:
            - $ref: '#/components/schemas/TwitchQuizPublicationStateDto'
          nullable: true
        errorCode: { type: string, nullable: true }
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
        code:
          type: string
          nullable: true
          description: Stable machine-readable error code.
          enum:
            - auth.api_client_header_required
            - role_administration.invalid_request
            - role_administration.user_not_found
            - role_administration.permanent_superadmin_protected
            - game_board.cell_not_found
            - game_board.active_team_required
            - game_board.active_team_no_active_game
            - game_board.active_team_not_found
            - game_board.active_team_not_confirmed
            - game_board.active_team_already_played
            - game_board.active_team_has_no_active_members
            - game_board.active_team_round_in_progress
            - game_board.team_played_state_no_active_game
            - game_board.team_played_state_not_found
            - game_board.team_played_state_not_confirmed
            - game_board.team_played_state_round_in_progress
            - game_setup.no_draft
            - game_setup.draft_exists
            - game_setup.invalid_title
            - game_setup.invalid_save_request
            - game_setup.cell_not_found
            - game_setup.cell_media_not_found
            - game_setup.invalid_cell_media_upload
            - game_setup.stale_version
            - game_lifecycle.draft_not_found
            - game_lifecycle.current_already_exists
            - game_lifecycle.active_already_exists
            - game_lifecycle.game_not_ready
            - game_lifecycle.game_not_active
            - game_lifecycle.registration_slots_required
            - game_lifecycle.questions_unavailable
            - game_lifecycle.invalid_team_size_limits
            - game_lifecycle.no_confirmed_teams
            - game_lifecycle.unconfirmed_teams
            - game_lifecycle.pending_invitations
            - game_lifecycle.pending_disband_requests
            - game_lifecycle.invalid_confirmed_team_roster
            - game_lifecycle.operation_failed
            - game_lifecycle.draft_delete_not_allowed
            - game_lifecycle.archive_not_allowed
            - game_lifecycle.game_not_found
            - game_finish.round_in_progress
            - game_finish.stale_version
            - game_finish.warnings_not_acknowledged
            - game_finish.modifier_state_invalid
            - game_finish.invalid_request
            - game_common.unexpected_server_error
            - game_common.too_many_requests
            - game_registration.not_open
            - game_registration.no_slots
            - game_registration.already_on_team
            - game_registration.team_not_found
            - game_registration.team_not_joinable
            - game_registration.team_roster_locked
            - game_registration.not_team_member
            - game_registration.invitation_invalid
            - game_registration.slot_not_found
            - game_registration.slot_not_available
            - game_registration.user_not_found
            - game_registration.pending_invitation
            - game_registration.pending_outgoing_invitation
            - game_registration.team_invite_not_allowed
            - game_registration.team_active_in_game
            - game_registration.team_already_played
            - game_registration.invalid_team_name
            - game_registration.team_name_taken
            - game_registration.team_name_required
            - game_registration.team_not_full
            - game_registration.disband_request_not_owned
            - game_registration.operation_failed
            - game_modifier.game_not_active
            - game_modifier.not_enabled
            - game_modifier.emergency_disabled
            - game_modifier_content_locked
            - game_modifier_revision_stale
            - game_modifier_compatibility_locked
            - game_modifier_archived
            - game_modifier_version_binding_missing
            - game_modifier.conflict_active
            - game_modifier.limit_reached
            - game_modifier.ordering_closed
            - game_modifier.active_team_member
            - game_modifier.insufficient_quiz_points
            - game_modifier.player_not_found
            - game_modifier.activation_not_found
            - game_modifier.activation_cancel_forbidden
            - game_modifier.activation_cancel_invalid_state
            - game_modifier.activation_cancel_reason_required
            - game_modifier.user_not_resolved
            - game_modifier.invalid_request
            - game_modifier_not_found
            - game_round.no_active_game
            - game_round.cell_not_found
            - game_round.cell_not_open
            - game_round.team_not_found
            - game_round.team_not_confirmed
            - game_round.team_has_no_active_members
            - game_round.awaiting_modifiers_required
            - game_round.already_in_progress
            - game_round.invalid_request
            - game_round.not_found
            - game_round.not_in_progress
            - game_round.stale_version
            - game_round.modifier_result_not_found
            - modifier_resolution.duplicate_group
            - modifier_resolution.duplicate_result
            - modifier_resolution.result_set_mismatch
            - modifier_resolution.group_set_mismatch
            - modifier_resolution.group_missing
            - modifier_resolution.group_members_mismatch
            - modifier_resolution.violation_comment_required
            - modifier_resolution.automatic_input_forbidden
            - modifier_resolution.boolean_required
            - modifier_resolution.non_negative_count_required
            - modifier_resolution.unsupported
            - modifier_resolution.missing
            - modifier_calculation.failed
            - behavior.invalid
            - behavior.rule_incompatible
            - formula.unsupported
            - formula.incompatible
            - resolution.invalid
            - round_facts.invalid
            - activation.duplicate
            - resolution.rule_status_required
            - resolution.automatic_required
            - resolution.boolean_required
            - resolution.non_negative_count_required
            - resolution.count_exceeds_resolved_kills
            - resolution.count_exceeds_activation_limit
            - resolution.per_activation_required
            - game_question.invalid_request
            - game_question.duplicate_code
            - game_question.not_found
            - game_question.category_not_found
            - game_question.category_not_empty
            - game_question.category_protected
            - game_question.import_invalid_fields
            - game_question.import_duplicate_code_in_file
            - game_question.import_category_unresolved
            - game_question.import_duplicate_code_existing
            - game_quiz.no_active_game
            - game_quiz.no_available_questions
            - game_quiz.answer_player_not_found
            - game_quiz.question_session_not_found
            - game_quiz.question_session_closed
            - game_quiz.already_answered
            - game_quiz.option_not_found
            - game_quiz.manual_award_player_not_found
            - game_quiz.manual_award_invalid_points
            - game_quiz.manual_award_invalid_operation
            - game_quiz.manual_award_invalid_reason
            - game_quiz.manual_award_insufficient_points
            - game_quiz.manual_award_duplicate_request_conflict
            - twitch_quiz.disabled
            - twitch_quiz.not_connected
            - twitch_quiz.publication_in_progress
            - twitch_quiz.pending_outcome
            - twitch_quiz.incompatible_question
            - twitch_quiz.question_message_too_long
            - twitch_quiz.options_message_too_long
            - twitch_quiz.result_message_too_long
        requestId:
          type: string
          nullable: true
          description: Server request correlation identifier for diagnostics.
...